Improved Sandbox Scene Handling

+++ sandbox/include/sandbox/panel/hierarchyPanel.h
@@ -14,12 +14,12 @@ public:
         auto* scene = m_Engine.getSceneManager().getActiveScene();
         const char* label = scene ? scene->name.c_str() : "No Scene";
 
-        if (ImGui::Button("Create Entity")) {
-            Entity* entity = m_Engine.createEntity("Entity");
-            m_Engine.moveToScene(entity);
+        if (ImGui::Button("Create Entity") && scene) {
+            Entity* entity = scene->createEntity("Entity");
+            m_SelectedEntity = entity;
         }
         ImGui::SameLine();
-        if (ImGui::Button("Duplicate Entity") && m_SelectedEntity) {
+        if (ImGui::Button("Duplicate Entity") && m_SelectedEntity && scene) {
             Entity* newEntity = scene->createEntity(m_SelectedEntity->name + " Copy");
             newEntity->transform = m_SelectedEntity->transform;
             for (const auto& [typeId, component] : m_SelectedEntity->getComponents()) {
@@ -27,11 +27,13 @@ public:
             }
         }
         ImGui::SameLine();
-        if (ImGui::Button("Delete Entity") && m_SelectedEntity) {
+        if (ImGui::Button("Delete Entity") && m_SelectedEntity && scene) {
             scene->destroyEntity(m_SelectedEntity);
             m_SelectedEntity = nullptr;
         }
+        ImGui::PushFont(ImGui::GetIO().Fonts->Fonts[1]);
         if (ImGui::TreeNodeEx(label, ImGuiTreeNodeFlags_DefaultOpen | ImGuiTreeNodeFlags_Framed | ImGuiTreeNodeFlags_Leaf | ImGuiTreeNodeFlags_SpanFullWidth)) {
+            ImGui::PopFont();
             if (scene) {
                 for (auto& entity : scene->getEntities()) {
                     bool selected = (m_SelectedEntity == entity.get());
